+++ src/main.rs
@@ -12,29 +12,37 @@ fn prepare_logging(
     console: bool,
     rotation: cli::LogRotation,
     retention: usize,
+    log_as: Option<&String>,
+    log_cmd_as: Option<&String>,
 ) -> Result<(), Box<dyn std::error::Error>> {
     let mut exe_dir = std::env::current_exe()?;
     exe_dir.pop();
 
+    let log_dir = match log_dir {
+        Some(log_dir) => log_dir.to_string(),
+        None => exe_dir.to_string_lossy().to_string(),
+    }
+    .replace("\\\\?\\", "");
+
+    let rotation = match rotation {
+        cli::LogRotation::Bytes(bytes) => flexi_logger::Criterion::Size(bytes),
+        cli::LogRotation::Daily => flexi_logger::Criterion::Age(flexi_logger::Age::Day),
+        cli::LogRotation::Hourly => flexi_logger::Criterion::Age(flexi_logger::Age::Hour),
+    };
+
     let mut logger = flexi_logger::Logger::try_with_env_or_str("debug")?
-        .log_to_file(
-            flexi_logger::FileSpec::default()
-                .directory(
-                    match log_dir {
-                        Some(log_dir) => log_dir.to_string(),
-                        None => exe_dir.to_string_lossy().to_string(),
-                    }
-                    .replace("\\\\?\\", ""),
-                )
-                .discriminant(format!("for_{}", name)),
-        )
+        .log_to_file({
+            let spec = flexi_logger::FileSpec::default().directory(log_dir.clone());
+
+            if let Some(log_as) = log_as {
+                spec.basename(log_as)
+            } else {
+                spec.discriminant(format!("for_{}", name))
+            }
+        })
         .append()
         .rotate(
-            match rotation {
-                cli::LogRotation::Bytes(bytes) => flexi_logger::Criterion::Size(bytes),
-                cli::LogRotation::Daily => flexi_logger::Criterion::Age(flexi_logger::Age::Day),
-                cli::LogRotation::Hourly => flexi_logger::Criterion::Age(flexi_logger::Age::Hour),
-            },
+            rotation,
             flexi_logger::Naming::Timestamps,
             flexi_logger::Cleanup::KeepLogFiles(retention),
         )
@@ -51,6 +59,27 @@ fn prepare_logging(
 
     if console {
         logger = logger.duplicate_to_stderr(flexi_logger::Duplicate::Info);
+    }
+
+    if let Some(log_cmd_as) = log_cmd_as {
+        logger = logger.add_writer(
+            "shawl-cmd",
+            Box::new(
+                flexi_logger::writers::FileLogWriter::builder(
+                    flexi_logger::FileSpec::default()
+                        .directory(log_dir)
+                        .basename(log_cmd_as),
+                )
+                .append()
+                .rotate(
+                    rotation,
+                    flexi_logger::Naming::Timestamps,
+                    flexi_logger::Cleanup::KeepLogFiles(retention),
+                )
+                .format(|w, _now, record| write!(w, "{}", &record.args()))
+                .try_build()?,
+            ),
+        );
     }
 
     logger.start()?;
@@ -78,6 +107,8 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
             console,
             common.log_rotate.unwrap_or_default(),
             common.log_retain.unwrap_or(2),
+            common.log_as.as_ref(),
+            common.log_cmd_as.as_ref(),
         )?;
     }